+++ app/src/main/java/com/chcreation/pointofsale/model/Payment.kt
@@ -4,7 +4,7 @@ import com.chcreation.pointofsale.EStatusCode
 
 data class Payment(
     var CREATED_DATE: String? = "",
-    var TOTAL_RECEIVED: Int? = 0,
+    var TOTAL_RECEIVED: Float? = 0F,
     var PAYMENT_METHOD: String? = "",
     var NOTE: String? = "All",
     var USER_CODE: String? = "",
