+++ app/src/main/java/com/chcreation/pointofsale/model/WholeSale.kt
@@ -3,8 +3,8 @@ package com.chcreation.pointofsale.model
 import com.chcreation.pointofsale.EStatusCode
 
 data class WholeSale(
-    var MIN_QTY: Int? = 0,
-    var MAX_QTY: Int? = 0,
-    var PRICE: Int? = 0,
+    var MIN_QTY: Float? = 0F,
+    var MAX_QTY: Float? = 0F,
+    var PRICE: Float? = 0F,
     var STATUS_CODE: String? = EStatusCode.ACTIVE.toString()
 )
